requirements.txt
python-telegram-bot==11.1.0
requests==2.22.0
tabulate==0.8.6
PyMySQL==0.9.3
Flask==1.1.1
PyTelegramBotAPI==2.2.3